pyproject.toml
[project]
name = "akkioTakeHomeProblem"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Peter", email = "peterfmutch@gmail.com"},
]
dependencies = [
    "Pillow>=10.4.0",
]
requires-python = "==3.9.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
